+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/domains/HyperSquareDomain.kt
@@ -16,8 +16,8 @@
 package space.kscience.kmath.domains
 
 import space.kscience.kmath.linear.Point
+import space.kscience.kmath.misc.UnstableKMathAPI
 import space.kscience.kmath.structures.Buffer
-import space.kscience.kmath.structures.RealBuffer
 import space.kscience.kmath.structures.indices
 
 /**
@@ -26,6 +26,7 @@ import space.kscience.kmath.structures.indices
  *
  * @author Alexander Nozik
  */
+@UnstableKMathAPI
 public class HyperSquareDomain(private val lower: Buffer<Double>, private val upper: Buffer<Double>) : RealDomain {
     public override val dimension: Int get() = lower.size
 
@@ -33,25 +34,9 @@ public class HyperSquareDomain(private val lower: Buffer<Double>, private val up
         point[i] in lower[i]..upper[i]
     }
 
-    public override fun getLowerBound(num: Int, point: Point<Double>): Double = lower[num]
-
     public override fun getLowerBound(num: Int): Double = lower[num]
 
-    public override fun getUpperBound(num: Int, point: Point<Double>): Double = upper[num]
-
     public override fun getUpperBound(num: Int): Double = upper[num]
-
-    public override fun nearestInDomain(point: Point<Double>): Point<Double> {
-        val res = DoubleArray(point.size) { i ->
-            when {
-                point[i] < lower[i] -> lower[i]
-                point[i] > upper[i] -> upper[i]
-                else -> point[i]
-            }
-        }
-
-        return RealBuffer(*res)
-    }
 
     public override fun volume(): Double {
         var res = 1.0