+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/domains/RealDomain.kt
@@ -15,45 +15,27 @@
  */
 package space.kscience.kmath.domains
 
-import space.kscience.kmath.linear.Point
+import space.kscience.kmath.misc.UnstableKMathAPI
 
 /**
  * n-dimensional volume
  *
  * @author Alexander Nozik
  */
+@UnstableKMathAPI
 public interface RealDomain : Domain<Double> {
-    public fun nearestInDomain(point: Point<Double>): Point<Double>
-
-    /**
-     * The lower edge for the domain going down from point
-     * @param num
-     * @param point
-     * @return
-     */
-    public fun getLowerBound(num: Int, point: Point<Double>): Double?
-
-    /**
-     * The upper edge of the domain going up from point
-     * @param num
-     * @param point
-     * @return
-     */
-    public fun getUpperBound(num: Int, point: Point<Double>): Double?
 
     /**
      * Global lower edge
-     * @param num
-     * @return
+     * @param num axis number
      */
-    public fun getLowerBound(num: Int): Double?
+    public fun getLowerBound(num: Int): Double
 
     /**
      * Global upper edge
-     * @param num
-     * @return
+     * @param num axis number
      */
-    public fun getUpperBound(num: Int): Double?
+    public fun getUpperBound(num: Int): Double
 
     /**
      * Hyper volume