+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/linear/MatrixWrapper.kt
@@ -1,20 +1,16 @@
 package space.kscience.kmath.linear
 
 import space.kscience.kmath.misc.UnstableKMathAPI
-import space.kscience.kmath.nd.Structure2D
 import space.kscience.kmath.nd.getFeature
 import space.kscience.kmath.operations.Ring
-import space.kscience.kmath.structures.asBuffer
-import kotlin.math.sqrt
 import kotlin.reflect.KClass
-import kotlin.reflect.safeCast
 
 /**
  * A [Matrix] that holds [MatrixFeature] objects.
  *
  * @param T the type of items.
  */
-public class MatrixWrapper<T : Any> internal  constructor(
+public class MatrixWrapper<T : Any> internal constructor(
     public val origin: Matrix<T>,
     public val features: Set<MatrixFeature>,
 ) : Matrix<T> by origin {
@@ -23,7 +19,8 @@ public class MatrixWrapper<T : Any> internal  constructor(
      * Get the first feature matching given class. Does not guarantee that matrix has only one feature matching the criteria
      */
     @UnstableKMathAPI
-    override fun <T : Any> getFeature(type: KClass<T>): T? = type.safeCast(features.find { type.isInstance(it) })
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : Any> getFeature(type: KClass<T>): T? = features.singleOrNull { type.isInstance(it) } as? T
         ?: origin.getFeature(type)
 
     override fun equals(other: Any?): Boolean = origin == other
@@ -61,29 +58,25 @@ public operator fun <T : Any> Matrix<T>.plus(newFeatures: Collection<MatrixFeatu
     }
 
 /**
- * Build a square matrix from given elements.
- */
-public fun <T : Any> Structure2D.Companion.square(vararg elements: T): Matrix<T> {
-    val size: Int = sqrt(elements.size.toDouble()).toInt()
-    require(size * size == elements.size) { "The number of elements ${elements.size} is not a full square" }
-    val buffer = elements.asBuffer()
-    return BufferMatrix(size, size, buffer)
-}
-
-/**
  * Diagonal matrix of ones. The matrix is virtual no actual matrix is created
  */
-public fun <T : Any, R : Ring<T>> GenericMatrixContext<T, R, *>.one(rows: Int, columns: Int): Matrix<T> =
-    VirtualMatrix(rows, columns) { i, j ->
-        if (i == j) elementContext.one else elementContext.zero
-    } + UnitFeature
+public fun <T : Any> LinearSpace<T, Ring<T>>.one(
+    rows: Int,
+    columns: Int,
+): Matrix<T> = VirtualMatrix(rows, columns) { i, j ->
+    if (i == j) elementAlgebra.one else elementAlgebra.zero
+} + UnitFeature
 
 
 /**
  * A virtual matrix of zeroes
  */
-public fun <T : Any, R : Ring<T>> GenericMatrixContext<T, R, *>.zero(rows: Int, columns: Int): Matrix<T> =
-    VirtualMatrix(rows, columns) { _, _ -> elementContext.zero } + ZeroFeature
+public fun <T : Any> LinearSpace<T, Ring<T>>.zero(
+    rows: Int,
+    columns: Int,
+): Matrix<T> = VirtualMatrix(rows, columns) { _, _ ->
+    elementAlgebra.zero
+} + ZeroFeature
 
 public class TransposedFeature<T : Any>(public val original: Matrix<T>) : MatrixFeature
 
@@ -91,9 +84,7 @@ public class TransposedFeature<T : Any>(public val original: Matrix<T>) : Matrix
  * Create a virtual transposed matrix without copying anything. `A.transpose().transpose() === A`
  */
 @OptIn(UnstableKMathAPI::class)
-public fun <T : Any> Matrix<T>.transpose(): Matrix<T> {
-    return getFeature<TransposedFeature<T>>()?.original ?: VirtualMatrix(
-        colNum,
-        rowNum,
-    ) { i, j -> get(j, i) } + TransposedFeature(this)
-}
+public fun <T : Any> Matrix<T>.transpose(): Matrix<T> = getFeature<TransposedFeature<T>>()?.original ?: VirtualMatrix(
+    colNum,
+    rowNum,
+) { i, j -> get(j, i) } + TransposedFeature(this)