+++ kmath-core/src/commonTest/kotlin/space/kscience/kmath/testutils/SpaceVerifier.kt
@@ -1,18 +1,18 @@
 package space.kscience.kmath.testutils
 
-import space.kscience.kmath.operations.Space
+import space.kscience.kmath.operations.Group
+import space.kscience.kmath.operations.ScaleOperations
 import space.kscience.kmath.operations.invoke
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
 
-internal open class SpaceVerifier<T>(
-    override val algebra: Space<T>,
+internal open class SpaceVerifier<T, out S>(
+    override val algebra: S,
     val a: T,
     val b: T,
     val c: T,
-    val x: Number
-) :
-    AlgebraicVerifier<T, Space<T>> {
+    val x: Number,
+) : AlgebraicVerifier<T, Group<T>> where S : Group<T>, S : ScaleOperations<T> {
     override fun verify() {
         algebra {
             assertEquals(a + b + c, a + (b + c), "Addition in $algebra is not associative.")